Catching up

+++ backends/p4test/run-p4-sample.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python3
 # Copyright 2013-present Barefoot Networks, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,8 +15,8 @@
 
 # Runs the compiler on a sample P4 V1.2 program
 
-from __future__ import print_function
-from subprocess import Popen,PIPE
+
+from subprocess import Popen, PIPE
 from threading import Thread
 import errno
 import sys
@@ -32,6 +32,7 @@ import glob
 SUCCESS = 0
 FAILURE = 1
 
+
 class Options(object):
     def __init__(self):
         self.binary = ""                # this program's name
@@ -43,7 +44,9 @@ class Options(object):
         self.dumpToJson = False
         self.compilerOptions = []
         self.runDebugger = False
+        self.runDebugger_skip = 0
         self.generateP4Runtime = False
+
 
 def usage(options):
     name = options.binary
@@ -58,9 +61,11 @@ def usage(options):
     print("          -a \"args\": pass args to the compiler")
     print("          --p4runtime: generate P4Info message in text format")
 
+
 def isError(p4filename):
     # True if the filename represents a p4 program that should fail
     return "_errors" in p4filename
+
 
 def ignoreStderr(options):
     for line in open(options.p4filename):
@@ -68,9 +73,11 @@ def ignoreStderr(options):
             return True
     return False
 
+
 class Local(object):
     # object to hold local vars accessable to nested functions
     pass
+
 
 def run_timeout(options, args, timeout, stderr):
     if options.verbose:
@@ -79,6 +86,7 @@ def run_timeout(options, args, timeout, stderr):
     local = Local()
     local.process = None
     local.filter = None
+
     def target():
         procstderr = None
         if stderr is not None:
@@ -92,7 +100,7 @@ def run_timeout(options, args, timeout, stderr):
             # available don't seem to actually work.
             local.filter = Popen(['sed', '-E',
                                   r's|^[-[:alnum:][:punct:][:space:]_/]*/([-[:alnum:][:punct:][:space:]_]+\.[ph]4?[:(][[:digit:]]+)|\1|'],
-                stdin=PIPE, stdout=outfile)
+                                 stdin=PIPE, stdout=outfile)
             procstderr = local.filter.stdin
         local.process = Popen(args, stderr=procstderr)
         local.process.wait()
@@ -115,10 +123,13 @@ def run_timeout(options, args, timeout, stderr):
         print("Exit code ", local.process.returncode)
     return local.process.returncode
 
+
 timeout = 10 * 60
 
-def compare_files(options, produced, expected):
-    if options.replace:
+
+def compare_files(options, produced, expected, ignore_case):
+    # p4info files should not change
+    if options.replace and "p4info" not in produced:
         if options.verbose:
             print("Saving new version of ", expected)
         shutil.copy2(produced, expected)
@@ -127,48 +138,66 @@ def compare_files(options, produced, expected):
     if options.verbose:
         print("Comparing", expected, "and", produced)
 
-    cmd = ("diff -B -u -w " + expected + " " + produced + " >&2")
+    args = "-B -u -w"
+    if ignore_case:
+        args = args + " -i"
+    cmd = ("diff " + args + " " + expected + " " + produced + " >&2")
     if options.verbose:
         print(cmd)
-    exitcode = subprocess.call(cmd, shell=True);
+    exitcode = subprocess.call(cmd, shell=True)
     if exitcode == 0:
         return SUCCESS
     else:
         return FAILURE
 
+
 def recompile_file(options, produced, mustBeIdentical):
     # Compile the generated file a second time
-    secondFile = produced + "-x";
+    secondFile = produced + "-x"
     args = ["./p4test", "-I.", "--pp", secondFile, "--std", "p4-16", produced] + \
-            options.compilerOptions
+        options.compilerOptions
+    if options.runDebugger:
+        if options.runDebugger_skip > 0:
+            options.runDebugger_skip = options.runDebugger_skip - 1
+        else:
+            args[0:0] = options.runDebugger.split()
+            os.execvp(args[0], args)
     result = run_timeout(options, args, timeout, None)
     if result != SUCCESS:
         return result
     if mustBeIdentical:
-        result = compare_files(options, produced, secondFile)
+        result = compare_files(options, produced, secondFile, False)
     return result
+
 
 def check_generated_files(options, tmpdir, expecteddir):
     files = os.listdir(tmpdir)
     for file in files:
-        if "p4info" in file:
-            continue
         if options.verbose:
             print("Checking", file)
         produced = tmpdir + "/" + file
         expected = expecteddir + "/" + file
-        if not os.path.isfile(expected):
+
+        if options.replace:
+            # Only create files when explicitly asked to do so
             if options.verbose:
                 print("Expected file does not exist; creating", expected)
             shutil.copy2(produced, expected)
-        else:
-            result = compare_files(options, produced, expected)
-            if result != SUCCESS and (file[-7:] != "-stderr" or not ignoreStderr(options)):
-                return result
+        elif not os.path.isfile(expected):
+            # The file is missing and we do not replace. This is an error.
+            print(
+                "Missing reference for file %s. Please rerun the test with the -f option turned on or rerun all tests using \"P4TEST_REPLACE=True make check\"." % expected)
+            return FAILURE
+        result = compare_files(
+            options, produced, expected, file[-7:] == "-stderr")
+        if result != SUCCESS and (file[-7:] != "-stderr" or not ignoreStderr(options)):
+            return result
     return SUCCESS
+
 
 def file_name(tmpfolder, base, suffix, ext):
     return tmpfolder + "/" + base + "-" + suffix + ext
+
 
 def process_file(options, argv):
     assert isinstance(options, Options)
@@ -177,6 +206,11 @@ def process_file(options, argv):
     basename = os.path.basename(options.p4filename)
     base, ext = os.path.splitext(basename)
     dirname = os.path.dirname(options.p4filename)
+    loops_unrolling = False
+    for option in options.compilerOptions:
+        if option == "--loopsUnroll":
+            loops_unrolling = True
+            break
     if "_samples/" in dirname:
         expected_dirname = dirname.replace("_samples/", "_samples_outputs/", 1)
     elif "_errors/" in dirname:
@@ -185,22 +219,25 @@ def process_file(options, argv):
         expected_dirname = dirname.replace("p4_14/", "p4_14_outputs/", 1)
     elif "p4_16/" in dirname:
         expected_dirname = dirname.replace("p4_16/", "p4_16_outputs/", 1)
+    elif loops_unrolling:
+        expected_dirname = dirname + "_outputs/parser-unroll"
     else:
         expected_dirname = dirname + "_outputs"  # expected outputs are here
     if not os.path.exists(expected_dirname):
         os.makedirs(expected_dirname)
 
     # We rely on the fact that these keys are in alphabetical order.
-    rename = { "FrontEndDump": "first",
-               "FrontEndLast": "frontend",
-               "MidEndLast": "midend" }
+    rename = {"FrontEndDump": "first",
+              "FrontEndLast": "frontend",
+              "MidEndLast": "midend"}
 
     if options.verbose:
         print("Writing temporary files into ", tmpdir)
     ppfile = tmpdir + "/" + basename                  # after parsing
-    referenceOutputs = ",".join(rename.keys())
+    referenceOutputs = ",".join(list(rename.keys()))
     stderr = tmpdir + "/" + basename + "-stderr"
-    p4runtimefile = tmpdir + "/" + basename + ".p4info.txt"
+    p4runtimeFile = tmpdir + "/" + basename + ".p4info.txt"
+    p4runtimeEntriesFile = tmpdir + "/" + basename + ".entries.txt"
 
     # Create the `json_outputs` directory if it doesn't already exist. There's a
     # race here since multiple tests may run this code in parallel, so we can't
@@ -220,12 +257,15 @@ def process_file(options, argv):
     def getArch(path):
         v1Pattern = re.compile('include.*v1model\.p4')
         psaPattern = re.compile('include.*psa\.p4')
-        with open(path, 'r') as f:
+        ubpfPattern = re.compile('include.*ubpf_model\.p4')
+        with open(path, 'r', encoding='utf-8') as f:
             for line in f:
                 if v1Pattern.search(line):
                     return "v1model"
                 elif psaPattern.search(line):
                     return "psa"
+                elif ubpfPattern.search(line):
+                    return "ubpf"
             return None
 
     if not os.path.isfile(options.p4filename):
@@ -236,22 +276,25 @@ def process_file(options, argv):
     if arch is not None:
         args.extend(["--arch", arch])
         if options.generateP4Runtime:
-            args.extend(["--p4runtime-format", "text"])
-            args.extend(["--p4runtime-file", p4runtimefile])
+            args.extend(["--p4runtime-files", p4runtimeFile])
+            args.extend(["--p4runtime-entries-files", p4runtimeEntriesFile])
 
     if "p4_14" in options.p4filename or "v1_samples" in options.p4filename:
-        args.extend(["--std", "p4-14"]);
+        args.extend(["--std", "p4-14"])
     args.extend(argv)
     if options.runDebugger:
-        args[0:0] = options.runDebugger.split()
-        os.execvp(args[0], args)
+        if options.runDebugger_skip > 0:
+            options.runDebugger_skip = options.runDebugger_skip - 1
+        else:
+            args[0:0] = options.runDebugger.split()
+            os.execvp(args[0], args)
     result = run_timeout(options, args, timeout, stderr)
 
     if result != SUCCESS:
         print("Error compiling")
-        print("".join(open(stderr).readlines()))
+        print(open(stderr).read())
         # If the compiler crashed fail the test
-        if 'Compiler Bug' in open(stderr).readlines():
+        if 'Compiler Bug' in open(stderr).read():
             return FAILURE
 
     expected_error = isError(options.p4filename)
@@ -266,9 +309,9 @@ def process_file(options, argv):
     lastFile = None
 
     for k in sorted(rename.keys()):
-        files = glob.glob(tmpdir + "/" + base + "*" + k + "*.p4");
+        files = glob.glob(tmpdir + "/" + base + "*" + k + "*.p4")
         if len(files) > 1:
-            print("Multiple files matching", k);
+            print("Multiple files matching", k)
         elif len(files) == 1:
             file = files[0]
             if os.path.isfile(file):
@@ -277,10 +320,10 @@ def process_file(options, argv):
                 lastFile = newName
 
     if (result == SUCCESS):
-        result = check_generated_files(options, tmpdir, expected_dirname);
+        result = check_generated_files(options, tmpdir, expected_dirname)
     if (result == SUCCESS) and (not expected_error):
         result = recompile_file(options, ppfile, False)
-    if (result == SUCCESS) and (not expected_error) and (lastFile is not None):
+    if (result == SUCCESS) and (not expected_error) and (lastFile is not None) and (arch is not "psa"):
         # Unfortunately compilation and pretty-printing of lastFile is
         # not idempotent: For example a constant such as 8s128 is
         # converted by the compiler to -8s128.
@@ -292,13 +335,15 @@ def process_file(options, argv):
         shutil.rmtree(tmpdir)
     return result
 
+
 def isdir(path):
     try:
         return stat.S_ISDIR(os.stat(path).st_mode)
     except OSError:
-        return False;
+        return False
 
-######################### main
+# main
+
 
 def main(argv):
     options = Options()
@@ -330,12 +375,14 @@ def main(argv):
                 usage(options)
                 sys.exit(FAILURE)
             else:
-                options.compilerOptions += argv[1].split();
+                options.compilerOptions += argv[1].split()
                 argv = argv[1:]
         elif argv[0][1] == 'D' or argv[0][1] == 'I' or argv[0][1] == 'T':
             options.compilerOptions.append(argv[0])
-        elif argv[0] == "-gdb":
+        elif argv[0][0:4] == "-gdb":
             options.runDebugger = "gdb --args"
+            if len(argv[0]) > 4:
+                options.runDebugger_skip = int(argv[0][4:]) - 1
         elif argv[0] == "--p4runtime":
             options.generateP4Runtime = True
         else:
@@ -347,10 +394,10 @@ def main(argv):
     if 'P4TEST_REPLACE' in os.environ:
         options.replace = True
 
-    options.p4filename=argv[-1]
+    options.p4filename = argv[-1]
     options.testName = None
     if options.p4filename.startswith(options.compilerSrcdir):
-        options.testName = options.p4filename[len(options.compilerSrcdir):];
+        options.testName = options.p4filename[len(options.compilerSrcdir):]
         if options.testName.startswith('/'):
             options.testName = options.testName[1:]
         if options.testName.endswith('.p4'):
@@ -361,6 +408,7 @@ def main(argv):
         print("Program was expected to fail")
 
     sys.exit(result)
+
 
 if __name__ == "__main__":
     main(sys.argv)
